+++ Kaenx/View/Import_old.xaml
@@ -1,0 +1,108 @@
+﻿<Page
+    x:Class="Kaenx.View.Import_old"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Kaenx.View"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:toolkit="using:Microsoft.Toolkit.Uwp.UI.Controls"
+    xmlns:extensions="using:Microsoft.Toolkit.Uwp.UI.Extensions"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <Grid.Background>
+            <SolidColorBrush Color="#FFE0E0E0" />
+        </Grid.Background>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="auto" />
+            <RowDefinition Height="auto" />
+            <RowDefinition Height="auto" />
+            <RowDefinition />
+            <RowDefinition Height="auto" />
+            <RowDefinition Height="auto" />
+        </Grid.RowDefinitions>
+
+        <StackPanel Padding="50 20 50 20" >
+            <TextBlock Text="{Binding ImportState}" FontSize="20" VerticalAlignment="Center" />
+            <TextBlock Text="{Binding ImportDevice}" FontSize="20" VerticalAlignment="Center" />
+        </StackPanel>
+
+
+        <ProgressBar Grid.Row="1" x:Name="ProgressMain" Maximum="6" Height="30" Foreground="Green" CornerRadius="0" />
+        <ProgressBar Grid.Row="2" x:Name="ProgressApp" Maximum="1" Value="1" Height="10" Foreground="SandyBrown" CornerRadius="0" />
+
+
+
+        <ListView x:Name="ViewDevicesList" Grid.Row="3" ItemsSource="{Binding DevicesList}" Background="white">
+            <ListView.ItemTemplate>
+                <DataTemplate>
+                    <Grid Margin="10">
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="50" />
+                            <ColumnDefinition />
+                        </Grid.ColumnDefinitions>
+                        <Grid.RowDefinitions>
+                            <RowDefinition />
+                            <RowDefinition />
+                        </Grid.RowDefinitions>
+
+                        <SymbolIcon Grid.RowSpan="2" MinWidth="10" Symbol="{Binding Icon}" Foreground="{Binding IconBrush}" />
+                        <TextBlock Grid.Column="1" Text="{Binding Name}" FontSize="18" />
+                        <TextBlock Grid.Row="1" Grid.Column="1" Text="{Binding Description}" Foreground="DarkBlue" VerticalAlignment="Center" />
+                    </Grid>
+                </DataTemplate>
+            </ListView.ItemTemplate>
+        </ListView>
+
+        <Border Grid.Row="3">
+            <Border.Background>
+                <SolidColorBrush Color="White" Opacity="0" />
+            </Border.Background>
+        </Border>
+
+
+        <StackPanel Grid.Row="4" VerticalAlignment="Center">
+            <Button x:Name="BtnBack" x:Uid="/Import/Back" HorizontalAlignment="Center" Margin="10" Click="ClickBack" IsEnabled="False"/>
+        </StackPanel>
+
+        <toolkit:Expander Grid.Row="5" ExpandDirection="Up" BorderBrush="Orange" BorderThickness="0 5 0 0" HorizontalContentAlignment="Stretch" >
+            <toolkit:Expander.Header>
+                <StackPanel  Orientation="Horizontal" HorizontalAlignment="Stretch">
+                    <TextBlock FontSize="18" Margin="20 0">
+                            <Run Text="Fehler:" />
+                            <Run Text="{Binding ImportError.Count}" />
+                    </TextBlock>
+
+                    <TextBlock FontSize="18" Margin="20 0">
+                        <Run Text="Warnungen:" />
+                        <Run Text="{Binding ImportWarning.Count}" />
+                    </TextBlock>
+                </StackPanel>
+            </toolkit:Expander.Header>
+
+            <Grid>
+                <Grid.Resources>
+                    <DataTemplate x:Key="ImportItemTemplate">
+                        <TextBlock Text="{Binding}" TextWrapping="WrapWholeWords" />
+                    </DataTemplate>
+                </Grid.Resources>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition />
+                    <ColumnDefinition />
+                </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="auto" />
+                    <RowDefinition />
+                </Grid.RowDefinitions>
+
+                <TextBlock Text="Fehler:" Margin="5" FontSize="16" />
+                <TextBlock Grid.Column="1" Text="Warnungen:" Margin="5" FontSize="16" />
+
+                <ListView Grid.Row="1" ItemsSource="{Binding ImportError}" ItemTemplate="{StaticResource ImportItemTemplate}" SelectionMode="None" />
+                <ListView Grid.Row="1" Grid.Column="1" ItemsSource="{Binding ImportWarning}" ItemTemplate="{StaticResource ImportItemTemplate}" SelectionMode="None" />
+            </Grid>
+        </toolkit:Expander>
+        
+    </Grid>
+</Page>
